servers/reddit_server/requirements.txt
# Reddit Server Requirements
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
websockets==12.0
python-dotenv==1.0.1
pydantic==2.5.3
praw==7.7.1
groq==0.4.1
